requirements.txt
python-dotenv
psutil
google-generativeai
cryptography
Pillow
moviepy
requests
